Move functions from single file to corresponding packages
