+++ turf/src/commonTest/kotlin/io/github/elcolto/geokjson/turf/measurement/DistanceTest.kt
@@ -1,0 +1,18 @@
+package io.github.elcolto.geokjson.turf.measurement
+
+import io.github.elcolto.geokjson.geojson.Position
+import io.github.elcolto.geokjson.turf.ExperimentalTurfApi
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+@ExperimentalTurfApi
+class DistanceTest {
+
+    @Test
+    fun testDistance() {
+        val a = Position(-73.67, 45.48)
+        val b = Position(-79.48, 43.68)
+
+        assertEquals(501.64563403765925, distance(a, b))
+    }
+}
